FINAL NOTEBOOK OF PROJECT
This is the file on which we should base our PowerPoint presentations. The instructions are clear: let's not just copy the code on it and simply present what we have done, but let's make it interesting, passionate, curious and also funny. Looking for images, gifs, being creative but also keeping in mind also to not detach too much from the content. I don't think it is the time in which we are required to show what we know or how good we are, but they expect something more, so let's surprise them!